Ended card api, not tested, created AllowedTo class to check if user is allowed to read or write from/to database

+++ src/main/kotlin/ru/kfd/bankan/bankanserver/controller/data/Cards.kt
@@ -6,13 +6,12 @@ import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import ru.kfd.bankan.bankanserver.payload.response.CardResponse
 import org.springframework.security.core.context.SecurityContextHolder
+import ru.kfd.bankan.bankanserver.controller.AllowedTo
 import ru.kfd.bankan.bankanserver.entity.CardEntity
 import ru.kfd.bankan.bankanserver.entity.ListToCardMappingEntity
 import ru.kfd.bankan.bankanserver.payload.request.CardCreationRequest
 import ru.kfd.bankan.bankanserver.payload.request.CardEditionRequest
 import ru.kfd.bankan.bankanserver.repository.*
-import java.time.LocalDate
-import java.util.*
 import ru.kfd.bankan.bankanserver.repository.ListToCardMappingRepository as ListToCardMappingRepository
 
 @RestController
@@ -21,32 +20,20 @@ class Cards(val cardRepository : CardRepository,
             val listRepository: ListRepository,
             val authInfoRepository: AuthInfoRepository,
             val listToCardMappingRepository: ListToCardMappingRepository,
-            val boardToListMappingRepository: BoardToListMappingRepository,
-            val boardToAssignedUserMappingRepository: BoardToAssignedUserMappingRepository,
+            val allowedTo: AllowedTo,
             val mapper : ObjectMapper){
-
-    fun allowed(listId : Int) : Optional<Boolean>
-    {
-        val login = SecurityContextHolder.getContext().authentication.principal
-        val creatorEntity = authInfoRepository.findByEmail(login.toString())
-            ?: return Optional.empty()
-        val creatorId = creatorEntity.userId
-        val boardListMappingEntity = boardToListMappingRepository.findByListId(listId)
-            ?: return Optional.empty()
-        val boardId = boardListMappingEntity.boardId
-        val users = boardToAssignedUserMappingRepository.getAllByBoardId(boardId)
-        if (users.find { it.userId == creatorId } == null) return Optional.of(false)
-        return Optional.of(true)
-    }
 
     // available for who?
     @GetMapping("/{cardId}")
     fun getCard(@PathVariable cardId : Int): ResponseEntity<String> {
-        // check if we are allowed to get a card
-        // the board can be open
-        // TODO
-
-
+        // check if card exists
+        val optionalCard = cardRepository.findById(cardId)
+        if(optionalCard.isEmpty) return ResponseEntity("Card with id $cardId does not exist", HttpStatus.NOT_FOUND)
+        // check if user have permission to get this card
+        val optionalBool = allowedTo.readByCardId(cardId)
+        if (optionalBool.isEmpty) return ResponseEntity("Something went wrong",HttpStatus.INTERNAL_SERVER_ERROR)
+        if (!optionalBool.get())
+            return ResponseEntity("You have no permissions to read this card", HttpStatus.FORBIDDEN)
         // get card
         val optional = cardRepository.findById(cardId)
         if(optional.isEmpty)
@@ -69,7 +56,7 @@ class Cards(val cardRepository : CardRepository,
         if (!listRepository.existsById(listId)) return ResponseEntity("There is no list with id $listId", HttpStatus.NOT_FOUND)
         // check if creator have permission to create a card
         if (!SecurityContextHolder.getContext().authentication.isAuthenticated) return ResponseEntity(HttpStatus.UNAUTHORIZED)
-        val optional = allowed(listId);
+        val optional = allowedTo.writeByListId(listId);
         if (optional.isEmpty) return ResponseEntity("Something went wrong",HttpStatus.INTERNAL_SERVER_ERROR)
         if (!optional.get())
             return ResponseEntity("You have no permissions to create a card in this list", HttpStatus.FORBIDDEN)
@@ -91,12 +78,18 @@ class Cards(val cardRepository : CardRepository,
         return ResponseEntity("Card created with id ${entity.id}", HttpStatus.OK)
     }
 
-    // available for who?
     @PatchMapping("/edit/{cardId}")
     fun editCard(@PathVariable cardId: Int, @RequestBody requestBody: CardEditionRequest) : ResponseEntity<String>{
-        val optional = cardRepository.findById(cardId)
-        if(optional.isEmpty) return ResponseEntity("Card with id $cardId does not exist", HttpStatus.NOT_FOUND)
-        val cardEntity = optional.get()
+        // check if card exists
+        val optionalCard = cardRepository.findById(cardId)
+        if(optionalCard.isEmpty) return ResponseEntity("Card with id $cardId does not exist", HttpStatus.NOT_FOUND)
+        // check if user have permission to change this card
+        val optionalBool = allowedTo.writeByCardId(cardId);
+        if (optionalBool.isEmpty) return ResponseEntity("Something went wrong",HttpStatus.INTERNAL_SERVER_ERROR)
+        if (!optionalBool.get())
+            return ResponseEntity("You have no permissions to edit this card", HttpStatus.FORBIDDEN)
+        // change the card
+        val cardEntity = optionalCard.get()
         if (requestBody.name != null) cardEntity.name = requestBody.name
         if (requestBody.cardContent != null) cardEntity.cardContent = requestBody.cardContent
         if (requestBody.changeColor) cardEntity.color = requestBody.color
@@ -104,9 +97,41 @@ class Cards(val cardRepository : CardRepository,
         return ResponseEntity("Card $cardId edited", HttpStatus.OK)
     }
 
+    @DeleteMapping("/{listId}/{cardId}")
+    fun deleteCardFromList(@PathVariable listId : Int, @PathVariable cardId : Int) : ResponseEntity<String>{
+        // check if card exists
+        if(cardRepository.findById(cardId).isEmpty)
+            return ResponseEntity("Card with id $cardId does not exist", HttpStatus.NOT_FOUND)
+        // check if card is not in this list
+        if(!listToCardMappingRepository.existsByListIdAndCardId(listId, cardId))
+            return ResponseEntity("Card with id $cardId is not in the list with id $listId", HttpStatus.NOT_FOUND)
+        // check if user have permission to change this card in the list
+        val optional = allowedTo.writeByListId(listId);
+        if (optional.isEmpty) return ResponseEntity("Something went wrong",HttpStatus.INTERNAL_SERVER_ERROR)
+        if (!optional.get())
+            return ResponseEntity("You have no permissions to delete this card from this list", HttpStatus.FORBIDDEN)
+        try {
+            // delete mapping of this card and this list
+            listToCardMappingRepository.deleteByListIdAndCardId(listId, cardId)
+        } catch (e : Exception){
+            return ResponseEntity("Unexpected error", HttpStatus.INTERNAL_SERVER_ERROR)
+        }
+        return ResponseEntity("Card with id $cardId deleted", HttpStatus.OK)
+    }
+
     @DeleteMapping("/{cardId}")
     fun deleteCard(@PathVariable cardId : Int) : ResponseEntity<String>{
+        // check if card exists
+        if(cardRepository.findById(cardId).isEmpty) return ResponseEntity("Card with id $cardId does not exist", HttpStatus.NOT_FOUND)
+        // check if user have permission to delete this card in all lists
+        val optional = allowedTo.writeByCardId(cardId);
+        if (optional.isEmpty) return ResponseEntity("Something went wrong",HttpStatus.INTERNAL_SERVER_ERROR)
+        if (!optional.get())
+            return ResponseEntity("You have no permissions to delete this card", HttpStatus.FORBIDDEN)
         try {
+            // delete all mappings with this card
+            listToCardMappingRepository.deleteAllByCardId(cardId)
+            // delete the card
             cardRepository.deleteById(cardId)
         } catch (e : Exception){
             return ResponseEntity("Unexpected error", HttpStatus.INTERNAL_SERVER_ERROR)
